ml/requirements.txt
Flask==2.3.3
numpy
opencv-python==4.8.1.78
tensorflow==2.17.1
pandas
gunicorn
pytesseract
Pillow
